multibagger_system/requirements.txt
pandas==2.1.0
numpy==1.24.3
yfinance==0.2.18
nsepy==0.8
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.3.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
ta==0.10.2
plotly==5.15.0
streamlit==1.25.0
dataclasses-json==0.5.14
typing-extensions==4.7.1
flask==2.3.0
flask-cors==4.0.0